+++ config/deps.ts
@@ -2,6 +2,7 @@ export {
     Application,
     Context,
     Router,
+    helpers,
     type RouteParams,
     type RouterContext,
     type RouterMiddleware,
@@ -16,3 +17,9 @@ export { Index } from "https://esm.sh/@upstash/vector";
 export { generateText } from "https://esm.sh/ai";
 export { createGoogleGenerativeAI } from "https://esm.sh/@ai-sdk/google";
 export { Filter } from "https://esm.sh/bad-words";
+
+// export {
+//   S3Client,
+//   HeadObjectCommand,
+//   PutObjectCommand,
+// } from "https://deno.land/x/aws_sdk/client-s3/mod.ts";